+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKBV69XZSBEAHX4Q0QT825CH"
+        "01KKC2FSWN8TH1J7Y53XANQN9Y"
     )
     r"""The unique identifier of the data source"""