+++ packages/orq-rc/RELEASES.md
@@ -6449,3 +6449,13 @@ Based on:
 - [python v4.5.0-rc.47] packages/orq-rc
 ### Releases
 - [PyPI v4.5.0-rc.47] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.47 - packages/orq-rc
+
+## 2026-03-10 14:32:59
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.750.0 (2.856.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.5.0-rc.48] packages/orq-rc
+### Releases
+- [PyPI v4.5.0-rc.48] https://pypi.org/project/orq-ai-sdk/4.5.0-rc.48 - packages/orq-rc